backend/kauri_chatbot_service/requirements.txt
# ============================================
# KAURI Chatbot Service - Python Dependencies
# ============================================

# Core Framework
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.38
alembic==1.15.1
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Vector Database
chromadb==0.4.24
pinecone-client==5.0.1

# LLM & Embeddings
openai==1.65.4
sentence-transformers==3.4.1
transformers==4.46.3
tokenizers==0.20.3
torch==2.6.0
accelerate==1.2.1

# RAG & Search
rank-bm25==0.2.2
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.42

# Document Processing
pypdf==5.3.1
pdfplumber==0.11.5
python-docx==1.1.2
beautifulsoup4==4.13.3

# Redis Cache
redis==5.2.1
hiredis==3.0.0

# HTTP Client
httpx==0.28.1
requests==2.32.3

# Authentication (pour validation JWT)
pyjwt==2.10.1
cryptography==44.0.2

# Monitoring & Logging
python-json-logger==3.2.1
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0

# Utilities
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
tqdm==4.67.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
